+++ sync/handler_blocks_request_test.go
@@ -35,7 +35,7 @@ func TestLatestBlocksRequestMessages(t *testing.T) {
 	tConfig.NodeNetwork = false
 	setup(t)
 
-	sid := util.RandInt(100)
+	sid := int(util.RandInt32(100))
 	pid := util.RandomPeerID()
 
 	t.Run("Reject request from unknown peers", func(t *testing.T) {
@@ -44,7 +44,7 @@ func TestLatestBlocksRequestMessages(t *testing.T) {
 
 		bdl := shouldPublishMessageWithThisType(t, tNetwork, message.MessageTypeBlocksResponse)
 		assert.Equal(t, bdl.Message.(*message.BlocksResponseMessage).ResponseCode, message.ResponseCodeRejected)
-		assert.Equal(t, bdl.Message.(*message.BlocksResponseMessage).From, 0)
+		assert.Equal(t, bdl.Message.(*message.BlocksResponseMessage).From, int32(0))
 	})
 
 	pub, _ := bls.GenerateTestKeyPair()
@@ -56,7 +56,7 @@ func TestLatestBlocksRequestMessages(t *testing.T) {
 
 		bdl := shouldPublishMessageWithThisType(t, tNetwork, message.MessageTypeBlocksResponse)
 		assert.Equal(t, bdl.Message.(*message.BlocksResponseMessage).ResponseCode, message.ResponseCodeRejected)
-		assert.Equal(t, bdl.Message.(*message.BlocksResponseMessage).From, 0)
+		assert.Equal(t, bdl.Message.(*message.BlocksResponseMessage).From, int32(0))
 	})
 
 	t.Run("NodeNetwork flag is not set", func(t *testing.T) {
@@ -68,7 +68,7 @@ func TestLatestBlocksRequestMessages(t *testing.T) {
 
 			bdl := shouldPublishMessageWithThisType(t, tNetwork, message.MessageTypeBlocksResponse)
 			assert.Equal(t, bdl.Message.(*message.BlocksResponseMessage).ResponseCode, message.ResponseCodeRejected)
-			assert.Equal(t, bdl.Message.(*message.BlocksResponseMessage).From, 0)
+			assert.Equal(t, bdl.Message.(*message.BlocksResponseMessage).From, int32(0))
 		})
 
 		t.Run("Accept request within `LatestBlockInterval`", func(t *testing.T) {
